Separate relationship and duplicate relationship suppliers
