+++ dhis-2/dhis-services/dhis-service-tracker/src/main/java/org/hisp/dhis/tracker/TrackerIdentifierCollector.java
@@ -189,7 +189,6 @@ public class TrackerIdentifierCollector
             if ( RelationshipPreheatKeySupport.hasRelationshipKey( relationship ) )
             {
                 RelationshipKey relationshipKey = RelationshipPreheatKeySupport.getRelationshipKey( relationship );
-                addIdentifier( identifiers, Relationship.class, relationshipKey.asString() );
 
                 addIdentifier( identifiers, TrackedEntity.class, relationshipKey.getFrom().getTrackedEntity() );
                 addIdentifier( identifiers, Enrollment.class, relationshipKey.getFrom().getEnrollment() );
